+++ anime_ml/__init__.py
@@ -1,9 +1,12 @@
+from .api.auth import authenticate
+from .api.dump import (ANIME_DETAILS_FILE, ANIME_LIST_FILE, FEATURES_FILENAME,
+                       read_jsonlines, write_jsonlines)
+from .api.request import get_anime, get_anime_list, get_profile
 from .cli import *  # file is empty
-from .extract.auth import authenticate
-from .extract.extract import *  # file is empty
-from .extract.request import *  # file is empty
-from .main import *  # file is empty
+from .main import download_anime_details, download_anime_list, main
 from .model.model import *  # file is empty
 from .model.transform import *  # file is empty
 
 from .settings import (BASE_URL, AUTH_URL, CLIENT_ID, CLIENT_SECRET)  # isort: skip
+
+__version__ = "0.0.4"