+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - 2020-12-26
+- added features in `models/features.pu`
+- added genres to `models/objects.py`
+- added methods in `main.py`
+
+## Changed
+- refined methods in `request.py` due to API documentation
+
 ## [0.0.3] - 2020-12-23
 ## Added
 - requests methods inside `anime_ml/api/request` to get user profile, animelist and generic details about an API